chore: rename features folder, update imports and add tsConfig improvements
