+++ src/components/AppProviders/AppProviders.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import { AdvocateProvider } from "@/feature-components/AdvocateListView/advocate-context";
+import { AdvocateProvider } from "@/features/AdvocateListView/advocate-context";
 
 export function AppProviders({ children }: { children: React.ReactNode }) {
   return <AdvocateProvider>{children}</AdvocateProvider>;
