+++ src/db/index.ts
@@ -1,7 +1,7 @@
 import 'dotenv/config';
 import { drizzle } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
-import * as schema from "./schema";
+import * as schema from "@/db/schema";
 
 const setup = () => {
   // for query purposes
